Add PREFIX_TEAM check and edit parse method in EditCommandParser class
Previously the parse method EditCommandParser class did not check for the team tag prefix `/T` when parsing an edit command, thus user was unable to edit the team tag of a person.

To fix this, we simply have the parse function tonkenize the team prefix when parsing the input, and also call the  `parseTagsForEdit` method for team tags.