Fix Edit Command to allow for editing teams
Previously the edit command could not edit team tags of a person. This problem arose as the `isAnyFieldEdited` method in the EditCommand class did not account for team tags. By simply adding the teams field as an argument within the method we have resolved the issue.

Additionally we also edited the equals method to check for team tags as well