Set fields to Private to match OOP
Previously the fields within classes such as Name, Email etc. are all public. This violates OOP principles as other classes have direct access to fields in another class.

In this commit all the fields have been changed to private, which cause issues as the code directly accesses the fields in other classes. To resolve this issue, the toString() function is called in place of the field directly. This also solves the issue of Optional fields down the road, as the toString method will resolve optionals