Edit PersonUtil class under tests
Edit person test has been failing as the PersonUtil did not account for the teams prefix.
To fix this, we added a chunk of the code which checks if the descriptor has team field. If it exists, the prefix is added along with the team tags.
Additionally a white space has to be appended after the PREFIX_TAG, such that the tokenize method in the EditCommandParser class does not consider the PREFIX_TEAM to be a tag to be added