Fix OOP for Tag
Changed the tagName field in Tag to be private. To get the tag name one has to call the to string method, which has also been modified to only give the tagName string without the square brackets. The removal of the square brackets in the toString function is made as
1. The toString method is not called elsewhere on any tag object,
2. The output of the toString method will be shown on the Ui, keeping the square brackets will result in them being shown to the user, which is not desireable.